+++ pistolwhip-mod/ThirdSpace_PistolWhip/Properties/AssemblyInfo.cs
@@ -1,0 +1,19 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("ThirdSpace_PistolWhip")]
+[assembly: AssemblyDescription("Third Space Vest haptic integration for Pistol Whip")]
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("Third Space")]
+[assembly: AssemblyProduct("ThirdSpace_PistolWhip")]
+[assembly: AssemblyCopyright("Copyright © 2024")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: ComVisible(false)]
+
+[assembly: Guid("d2e3f4a5-b6c7-8901-def2-234567890123")]
+
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
+